bot/pyproject.toml
[project]
name = "expanses_tracker"
version = "0.1.0"
description = "Expanses tracker telegram bot"
authors = [
    {name = "Alessandro Giacche",email = "alessandrogiacche@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot (>=22.3,<23.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "SQLAlchemy (>=2.0.0,<3.0.0)",
    "psycopg2-binary (>=2.9.0,<3.0.0)",
    "mysqlclient (>=2.2.0,<3.0.0)",
    "alembic (>=1.12.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "*", from = "expanses_tracker"}]

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.8"
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
